Refactored project structure: moved Profile and Wardrobe screens to their own packages.
