+++ app/src/main/java/com/drape/navigation/HomeNavGraph.kt
@@ -4,10 +4,10 @@ import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.composable
 import androidx.navigation.navigation
-import com.drape.ui.wardrobe.WardrobeScreen
+import com.drape.ui.outfit_creator.WardrobeScreen
 import com.drape.ui.home.CamerinoScreen
 import com.drape.ui.home.HomeScreen
-import com.drape.ui.home.ProfileScreen
+import com.drape.ui.profile.ProfileScreen
 import com.drape.ui.upload_clothes.UploadItemScreen
 
 /**
